# specs/001-control-plane-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Control Plane API
  description: Multi-tenant SaaS control plane for organization, tenant, and subscription management
  version: 1.0.0
  contact:
    name: Control Plane Team
servers:
  - url: https://control-plane.example.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
  /auth/login:
    get:
      tags: [Auth]
      summary: Initiate OIDC login
      description: Redirects user to Zitadel for authentication
      operationId: login
      security: []
      parameters:
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
          description: Optional override for redirect URI
      responses:
        '302':
          description: Redirect to Zitadel authorization endpoint
          headers:
            Location:
              schema:
                type: string
        '400':
          description: Invalid request
  /auth/callback:
    get:
      tags: [Auth]
      summary: OIDC callback handler
      description: Handles callback from Zitadel after authentication
      operationId: callback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Zitadel
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid callback request
        '401':
          description: Authentication failed
        '500':
          description: Server error during authentication
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns information about the authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        last_login_at:
          type: string
          format: date-time
          nullable: true
...